requirements.txt
fastapi[standard]==0.115.4
requests
pydantic
apscheduler
uvicorn
PyGithub
python-dotenv
